GameManager 의존성 제거
- 불필요한 참조 삭제

+++ Assets/Scripts/GameSystem/OldGameSystem.cs
@@ -77,7 +77,7 @@ public class OldGameSystem : Singleton<OldGameSystem>
         // DataLoader.SavePlayerData(saveList);
 
         // 해당 날짜 설정
-        DayData = DataLoader.GetDayData(date);
+        DayData = DataLoader.GetDayData((uint)date);
         dateIndex = date;
         currentLocation = DayData.startLocation;
         isScreenOn = false;
